Refactoring duplication into abstract response class and renaming Card and Customer message classes
