+++ src/Message/AbstractRequest.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Omnipay\Pin\Message;
+
+abstract class AbstractRequest extends \Omnipay\Common\Message\AbstractRequest
+{
+    protected $liveEndpoint = 'https://api.pin.net.au/1';
+    protected $testEndpoint = 'https://test-api.pin.net.au/1';
+
+    public function getSecretKey()
+    {
+        return $this->getParameter('secretKey');
+    }
+
+    public function setSecretKey($value)
+    {
+        return $this->setParameter('secretKey', $value);
+    }
+
+    public function getEndpoint()
+    {
+        return $this->getTestMode() ? $this->testEndpoint : $this->liveEndpoint;
+    }
+
+    public function sendRequest($action, $data = null)
+    {
+        // don't throw exceptions for 4xx errors
+        $this->httpClient->getEventDispatcher()->addListener(
+            'request.error',
+            function ($event) {
+                if ($event['response']->isClientError()) {
+                    $event->stopPropagation();
+                }
+            }
+        );
+
+        $httpResponse = $this->httpClient->post($this->getEndpoint() . $action, null, $data)
+            ->setHeader('Authorization', 'Basic ' . base64_encode($this->getSecretKey() . ':'));
+
+        return $httpResponse->send();
+    }
+}
