--- src/Message/CardRequest.php
@@ -1,68 +1,0 @@
-<?php
-
-namespace Omnipay\Pin\Message;
-
-use Omnipay\Common\Message\AbstractRequest;
-
-/**
- * Pin Purchase Request
- */
-class CardRequest extends AbstractRequest
-{
-    protected $liveEndpoint = 'https://api.pin.net.au/1';
-    protected $testEndpoint = 'https://test-api.pin.net.au/1';
-
-    public function getSecretKey()
-    {
-        return $this->getParameter('secretKey');
-    }
-
-    public function setSecretKey($value)
-    {
-        return $this->setParameter('secretKey', $value);
-    }
-
-    public function getData()
-    {
-        $data = array();
-        $this->getCard()->validate();
-
-        $data['number'] = $this->getCard()->getNumber();
-        $data['expiry_month'] = $this->getCard()->getExpiryMonth();
-        $data['expiry_year'] = $this->getCard()->getExpiryYear();
-        $data['cvc'] = $this->getCard()->getCvv();
-        $data['name'] = $this->getCard()->getName();
-        $data['address_line1'] = $this->getCard()->getAddress1();
-        $data['address_line2'] = $this->getCard()->getAddress2();
-        $data['address_city'] = $this->getCard()->getCity();
-        $data['address_postcode'] = $this->getCard()->getPostcode();
-        $data['address_state'] = $this->getCard()->getState();
-        $data['address_country'] = $this->getCard()->getCountry();
-
-        return $data;
-    }
-
-    public function sendData($data)
-    {
-        // don't throw exceptions for 4xx errors
-        $this->httpClient->getEventDispatcher()->addListener(
-            'request.error',
-            function ($event) {
-                if ($event['response']->isClientError()) {
-                    $event->stopPropagation();
-                }
-            }
-        );
-
-        $httpResponse = $this->httpClient->post($this->getEndpoint() . '/cards', null, $data)
-            ->setHeader('Authorization', 'Basic ' . base64_encode($this->getSecretKey() . ':'))
-            ->send();
-
-        return $this->response = new Response($this, $httpResponse->json());
-    }
-
-    public function getEndpoint()
-    {
-        return $this->getTestMode() ? $this->testEndpoint : $this->liveEndpoint;
-    }
-}
